test: move MockXdsConfigTracker and MockXdsResourcesDelegate to shared mocks
Move MockXdsConfigTracker and MockXdsResourcesDelegate from local test
files to test/mocks/config/mocks.h to allow reuse and avoid duplication.
Update BUILD dependencies accordingly.

TAG=agy
CONV=b942214e-2f01-4571-9478-b3eda0877a4c
